requirements.txt
vulkan
glfw
numpy
pyrr
pillow
dearpygui
GPUtil
pynvml
psutil
PyOpenGL
curses
winsound